backend/pyproject.toml
[project]
name = "iris-tailor-backend"
version = "0.1.0"
description = "Backend for Iris Tailor — sewing pattern adjustment via AI"
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "lxml>=5.1.0",
    "numpy>=1.26.0",
    "anthropic>=0.40,<1",
    "python-multipart>=0.0.9",
    "pillow>=10.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "trimesh>=4.12.1",
    "replicate>=1.0.7",
    "fpdf2>=2.7.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "black>=24.2.0",
    "mypy>=1.8.0",
    "hypothesis>=6.98.0",
    "pygltflib>=1.16.5",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
markers = [
    "integration: hits real external APIs; requires credentials (ANTHROPIC_API_KEY or REPLICATE_API_TOKEN)",
]

[tool.mypy]
python_version = "3.11"
strict = true
